Insert an element instead of a node in a list.

+++ src/main/java/io/github/harperkej/App.java
@@ -18,7 +18,7 @@ public class App {
 
         List<Node<Integer>> nodes = new ArrayList<>();
         for (int i = 0; i < 10; i++) {
-            Node node = new Node(i);
+            Node node = new Node((long) i);
             node.setObject(i);
             nodes.add(node);
         }
